feat(k8s-ui): extract ResourcesSidebar as standalone component
