+++ internal/version/birthtime_other.go
@@ -1,0 +1,7 @@
+//go:build !darwin && !linux && !windows
+
+package version
+
+func dirBirthtime(_ string) int64 {
+	return 0
+}
