+++ packages/k8s-ui/src/components/resources/index.ts
@@ -15,3 +15,5 @@ export * from './resource-utils-traefik'
 export * from './resource-utils-velero'
 export { ResourcesView, ResourcesViewDataContext } from './ResourcesView'
 export type { ResourceQueryResult } from './ResourcesView'
+export { ResourcesSidebar } from './ResourcesSidebar'
+export type { ResourcesSidebarProps, SelectedKindInfo, PinnedItem } from './ResourcesSidebar'
